Add locked-only download option.
Add an option to download only locked files.  This drove a lot of code cleanups
that were not strictly necessary for the feature.

+++ downloadImages.py
@@ -3,16 +3,13 @@
 '''
 downloadImages -- Download images from a DCF volume such as an SD card.
 
-It defines classes_and_methods
-
 @author:     John Dykstra
 
 @copyright:  2017-2023 John Dykstra. All rights reserved.
 
-@license:    Apache
+@license:    MIT
 
 @contact:    jdykstra72@gmail.com
-@deffield    updated: Updated
 
 ??  Todo - VT100 control sequences not being interpreted on Windows.
                 See https://stackoverflow.com/questions/51680709/colored-text-output-in-powershell-console-using-ansi-vt100-codes
@@ -20,8 +17,11 @@ It defines classes_and_methods
            Get info via dialog
    
 '''
-from __future__ import division
-from __future__ import print_function
+
+''' 
+    We try to be filesystem-case-sensitive-agnostic.  The case of source filenames and extensions is preserved in
+    the destination directory, but we don't allow multiple destination files that differ only in case.
+'''
 
 from builtins import zip
 from builtins import str
@@ -39,12 +39,11 @@ import traceback
 
 from argparse import ArgumentParser
 from argparse import RawDescriptionHelpFormatter
-#  from __builtin__ import True
 
 __all__ = []
-__version__ = 1.7
+__version__ = 1.8
 __date__ = '2017-04-06'
-__updated__ = '2023-05-11'
+__updated__ = '2023-07-09'
 
 DEBUG = 1
 TESTRUN = 0
@@ -98,17 +97,36 @@ def createDestinationDir(destPath, name):
          os.makedirs(d)
     
     return d
+
+# Class representing a single image.  That image may have more than one associated image file, such as a JPEG and a RAW file.
+class Image:
+    def __init__(self, srcFilename, srcPath, extension, fileLocked, dstFilename):
+        self.srcFilename = srcFilename
+        self.srcPath = srcPath
+        self.extensions = [extension]
+        self.fileLocked = fileLocked
+        self.dstFilename = dstFilename
     
-# Return a dictionary describing all of the image files on the source..
-def findSourceImages(src):
+    def addFileExtension(self, extension):
+        self.extensions.append(extension)
+
+    def __str__(self):
+        return "ImageFile:  filename = {0}, srcPath = {1}, destPath = {2}, extensions = {3}, skip = {4}, duplicate = {5}, xmp = {6}, xmpPath = {7}, xmpName = {8}, xmpDestPath = {9}, xmpDestName = {10}, xmpDestFullPath = {11}".format(self.name, self.srcPath, self.destPath, self.extensions, self.skip, self.duplicate, self.xmp, self.xmpPath, self.xmpName, self.xmpDestPath, self.xmpDestName, self.xmpDestFullPath)
+        
+    def __repr__(self):
+        return self.__str__()
+    
+# Return a dictionary describing all of the image files on the source, indexed by the image name.
+def findSourceImages(src, downloadLockedOnly):
     images = {}
     jpegCnt = 0
     movCnt = 0
+    lockedFileCnt = 0
     nearRollover = False
     rolloverOccurred = False
     
     # Enumerate the image files on the source volume.
-    for dirpath, dirs, files in os.walk(src):
+    for dirpath, _, files in os.walk(src):
         for f in files:
             
             # Ignore files that are unlikely to be camera image files,
@@ -118,33 +136,52 @@ def findSourceImages(src):
             fparts = f.split(".")
             if len(fparts) != 2:
                 continue
-            imgName = fparts[0].upper()
-            cleanImgName = imgName.replace("_", "")
-            extension = fparts[-1].upper()
-            if extension not in ['JPG', 'NEF', 'MOV', 'MP4']:
+            srcFilename = fparts[0]
+            extension = fparts[-1]
+            dstFilename = srcFilename.replace("_", "")  # Remove underscores used by Nikon
+            imageName = dstFilename.upper()  
+            if extension.upper() not in ['JPG', 'NEF', 'MOV', 'MP4']:
                 continue
             
-            # Remember if the number part of the image name is getting near the rollover point.
-            nearRollover |= cleanImgName[-4] == '9'
-            rolloverOccurred |= cleanImgName[-4:] == "9999"
-            
-            # Dictionary "images" is indexed by the image name.  Its entries are themselves
-            # dictionaries, containing the filename, the path to that filename, and
-            # a list of the file extensions seen for that filename.
-            if cleanImgName in images:
-                images[cleanImgName]["extensions"].append(extension)
+            # If write protect was set on an image by the camera, it will appear on
+            # MacOS as the user-immutable flag.  FWIW, this flag can be seen using
+            # "ls -lhdO".  On Windows, we just look for read-only.
+            srcFullPath = os.path.join(dirpath, srcFilename + "." + extension)
+            if 'darwin' in sys.platform:
+                fileLocked = os.stat(srcFullPath).st_flags & stat.UF_IMMUTABLE
             else:
-                images[cleanImgName] = dict(extensions = [extension], srcPath = dirpath, origName = imgName)
+                fileLocked = not os.access(srcFullPath, os.W_OK)
+
+            # Remember if the number part of the image name is getting near the rollover point.
+            nearRollover |= imageName[-4] == '9'
+            rolloverOccurred |= imageName[-4:] == "9999"
+
+            # If we're downloading only locked images, ignore all the rest.
+            if downloadLockedOnly and not fileLocked:
+                continue
+            
+            # Dictionary "images" is indexed by the image name, forced to uppercase.  
+            if imageName in images:
+                images[imageName].addExtension(extension)
+            else:
+                images[imageName] = Image(srcFilename, dirpath, extension, fileLocked, dstFilename)
                         
-            if extension == 'JPG':
+            if extension.upper() == 'JPG':
                 jpegCnt += 1
-            elif extension == 'MOV':
+            elif extension.upper() == 'MOV':
                 movCnt += 1
                 
+            if fileLocked:
+                lockedFileCnt += 1
+
     if jpegCnt > 0:
         print("WARNING:  {0} JPEG files found!".format(jpegCnt))
     if movCnt > 0:
         print("{0} video files found.".format(movCnt))
+    if lockedFileCnt > 0:
+        print("{0} files are locked.".format(lockedFileCnt))         
+    elif downloadLockedOnly:
+        print("WARNING:  Downloading locked files only, but no locked files found.")
     if rolloverOccurred:
         print("WARNING:  Image numbers rolled over!")
     elif nearRollover:
@@ -152,77 +189,70 @@ def findSourceImages(src):
     
     return images
 
-# Return a list of files already in the destination directory.
+# Return a list of files already in a destination directory.
 # ?? Too complex.  Either set a "skip" key in the per-image dictionary, or delete the per-image dictionary from
 # ?? the images dictionary.
 # ?? Except a file might be a duplicate in one destination directory, and not another.
 def lookForDuplicates(images, dst):
     duplicates =[]
     
-    for cleanImgName in iter(images):
-        for ext in images[cleanImgName]["extensions"]:
-            filename = cleanImgName + "." + ext
-            dstpath = os.path.join(dst, filename)
-            if os.path.exists(dstpath):
-                srcpath = os.path.join(images[cleanImgName]["srcPath"], images[cleanImgName]["origName"] + "." + ext)
-                if (os.stat(dstpath).st_size == os.stat(srcpath).st_size):
-                    duplicates.append(cleanImgName)
+    for imageName in iter(images):
+        for ext in images[imageName].extensions:
+            dstFullPath = os.path.join(dst, images[imageName].dstFilename + "." + ext)
+            if os.path.exists(dstFullPath):
+                srcFullPath = os.path.join(images[imageName].srcPath, images[imageName].srcFilename + "." + ext)
+                if (os.stat(dstFullPath).st_size == os.stat(srcFullPath).st_size):
+                    duplicates.append(imageName)
     
     return duplicates
     
 
-# Copy the image files from the source to the destination and create the XMP sidecar file.
-def copyImageFiles(images, destinationDirs, skips, description, delete=False):
+# Copy the image files from the source to the destination and create a XMP sidecar file for each.
+def copyImageFiles(images, destinationDirs, skips, description, downloadLockedOnly=False, delete=False):
 
     progress = 0
-    for name in iter(images):
-        entry = images[name]
+    for imageName in iter(images):
+        image = images[imageName]
         progress += 1
         for dest, skip in zip(destinationDirs, skips):
-            if name not in skip:
-                writeProtect = False
-                for ext in entry["extensions"]:                    
-                    srcpath = os.path.join(entry["srcPath"], entry["origName"] + "." + ext)
-                    dstpath = os.path.join(dest, name + "." + ext)
-                    sys.stdout.write("{0}%:  {1} to {2}.{3}\r".format(int((progress * 100)/len(images)), name, dstpath, cleol))
+            if imageName not in skip:
+                for ext in image.extensions:                    
+                    srcFullpath = os.path.join(image.srcPath, image.srcFilename + "." + ext)
+                    dstFullPath = os.path.join(dest, imageName + "." + ext)
+                    sys.stdout.write("{0}%:  {1} to {2}.{3}\r".format(int((progress * 100)/len(images)), imageName, dstFullPath, cleol))
                     sys.stdout.flush()
 
-                    # If write protect was set on an image by the camera, it will appear on
-                    # MacOS as the user-immutable flag.  FWIW, this flag can be seen using
-                    # "ls -lhdO".  On Windows, we just look for read-only.
-                    if 'darwin' in sys.platform:
-                        writeProtect = os.stat(srcpath).st_flags & stat.UF_IMMUTABLE 
-                    else:
-                        writeProtect = not os.access(srcpath, os.W_OK)
-
-                    # Copy the image file.
-                    shutil.copy2(srcpath, dstpath)
+                    # Copy the image file.  If we're only copying locked files, skip unlocked files.
+                    if not downloadLockedOnly or image.fileLocked:
+                        shutil.copy2(srcFullpath, dstFullPath)
                 
                     # If write protect was set on the source file, clear it on the destination.  We'll
                     # treat it specially below when we create the XMP sidecar file.  If we're going
                     # to delete the source file, also clear write protect on it.
-                    if writeProtect:
+                    if image.fileLocked:
                         if 'darwin' in sys.platform:
-                            os.chflags(dstpath, os.stat(dstpath).st_flags & ~stat.UF_IMMUTABLE)
+                            os.chflags(dstFullPath, os.stat(dstFullPath).st_flags & ~stat.UF_IMMUTABLE)
                             if delete:
-                                os.chflags(srcpath, os.stat(srcpath).st_flags & ~stat.UF_IMMUTABLE)
+                                os.chflags(srcFullpath, os.stat(srcFullpath).st_flags & ~stat.UF_IMMUTABLE)
                         else:
-                            os.chmod(dstpath, stat.S_IWRITE)
+                            os.chmod(dstFullPath, stat.S_IWRITE)
                             if delete:
-                                os.chmod(srcpath, stat.S_IWRITE)
+                                os.chmod(srcFullpath, stat.S_IWRITE)
 
                 # Create the sidecar file.
                 # ?? Use multi-line string constant?
                 # ?? The write protect part could be coded as:
-                # ??      writeProtect and "Purple" or "None"
-                sidecar = open(os.path.join(dest, name+".xmp"), "w")
+                # ??      fileLocked and "Purple" or "None"
+                if image.fileLocked:
+                    print("{0} is {1}immutable".format(imageName, "" if image.fileLocked else "not "))
+                sidecar = open(os.path.join(dest, image.dstFilename + ".xmp"), "w")
                 sidecar.write("<x:xmpmeta xmlns:x=\"adobe:ns:meta/\">\n")
                 sidecar.write("<rdf:RDF xmlns:rdf=\"http://www.w3.org/1999/02/22-rdf-syntax-ns#\">\n")
                 sidecar.write("\n")
                 sidecar.write("  <rdf:Description rdf:about=\"\"\n")
                 sidecar.write("     xmlns:xmp=\"http://ns.adobe.com/xap/1.0/\"\n")
                 sidecar.write("     xmlns:dc=\"http://purl.org/dc/elements/1.1/\"\n")
-                if writeProtect:
+                if image.fileLocked:
                     sidecar.write("     xmp:Label=\"Purple\"\n")
                 sidecar.write("     >\n")
                 sidecar.write("     <dc:description>\n")
@@ -240,7 +270,7 @@ def copyImageFiles(images, destinationDirs, skips, description, delete=False):
     sys.stdout.flush()      
             
 # Programmatic API to this module.  Returns name (not path) of destination directories.
-def doDownload(destinationPaths, tag, description, delete=False, verbose=False):
+def doDownload(destinationPaths, tag, description, downloadLockedOnly=False, delete=False, verbose=False):
        
     # Find the source volume.  We can only handle one.
     sourceVols = findSourceVolume()
@@ -251,14 +281,14 @@ def doDownload(destinationPaths, tag, description, delete=False, verbose=False):
     sourceVol=sourceVols[0]
 
     # Find image files on the source volume.
-    images = findSourceImages(sourceVol[1])
+    images = findSourceImages(sourceVol[1], downloadLockedOnly)
     print("{0} image files found on {1}.".format(len(images), sourceVol[0]))
     
     # If we're supposed to delete the source images, make sure that we can.
     if (delete and not os.access(sourceVol[1], os.W_OK)):
         raise CLIError
     
-    # Handle multiple possible destinations.
+    # Look for existing duplicates on the destination volumes.
     # DestinationDirs and duplicates are lists in the same order as the
     # entries in destinationPaths.
     destinationDirs = []
@@ -281,7 +311,10 @@ def doDownload(destinationPaths, tag, description, delete=False, verbose=False):
     # ?? Having matching tuples of destination directories and duplicate lists seems 
     # ?? unnecessarily complex.  Why not call copyImageFiles() once for each destination
     # ?? directory?
-    copyImageFiles(images, destinationDirs, duplicates, description, delete)
+    # ?? We could also pass in the destination directory and duplicate list as a tuple.
+    # On the other hand, this enables us to write all the destinations while each source
+    # file is still open and in cache.
+    copyImageFiles(images, destinationDirs, duplicates, description, downloadLockedOnly, delete)
      
     # Delete the source files.
     if delete:
@@ -340,6 +373,7 @@ USAGE
         parser.add_argument("-v", "--verbose", dest="verbose", action="count", default=0, help="set verbosity level [default: %(default)s]")
         parser.add_argument("-t", "--tag", dest="tag", default="Downloaded Images", help="Tag used as destination directory name. [default: %(default)s]" )
         parser.add_argument("-d", "--description", dest="description", help="Description saved in each photo's sidecar.")
+        parser.add_argument("-L", "--locked-only", dest="downloadLockedOnly", action='store_true', help="Only download locked files.")
         parser.add_argument("-D", "--delete", dest="delete", action='store_true', help="Delete files from card after successful download.")
         parser.add_argument("-a", "--automate", dest="automate", action='store_true', help="Execute deleteUnderscore and Photos.")
         parser.add_argument('-V', '--version', action='version', version=program_version_message)
@@ -358,11 +392,14 @@ USAGE
             if not os.path.exists(path):
                 print("Error:  Destination path \"{0}\" doesn't exist.".format(path))
                 return 2
+        if args.downloadLockedOnly and args.delete:
+            print("Error:  Delete and locked-only options are mutually exclusive.")
+            return 2
     
         if 'darwin' in sys.platform:
             caffeinateProcess = subprocess.Popen(('caffeinate', '-i'))
 
-        dirName = doDownload(args.destinations, args.tag, args.description, args.delete, args.verbose)
+        dirName = doDownload(args.destinations, args.tag, args.description, args.downloadLockedOnly, args.delete, args.verbose)
         
         if caffeinateProcess != None:
             caffeinateProcess.terminate()
